# Conflicts:
#	app/src/main/java/a75f/io/renatus/FloorPlanFragment.java

+++ haystack/src/main/java/a75f/io/api/haystack/CCUTagsDb.java
@@ -1160,7 +1160,7 @@ public class CCUTagsDb extends HServer {
     //Delete all the hisItem entries older than 24 hrs.
     public void removeExpiredHisItems(HRef id) {
         HDict entity = readById(id);
-        double backfillduration = CCUHsApi.getInstance().readPointPriorityValByQuery("point and backfill and duration");
+        long backfillduration =  CCUHsApi.getInstance().readPointPriorityValByQuery("point and backfill and duration").intValue();
         QueryBuilder<HisItem> hisQuery = hisBox.query();
         hisQuery.equal(HisItem_.rec, entity.get("id").toString())
                 .less(HisItem_.date, System.currentTimeMillis() - backfillduration*60*60*1000)
